+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 [Cherry](https://github.com/squint-cljs/cherry): Experimental ClojureScript to ES6 module compiler
 
+## Unreleased
+
+- [#71](https://github.com/squint-cljs/cherry/issues/71): support alias with dashes
+- [#77](https://github.com/squint-cljs/cherry/issues/77): support async/await in variadic functions
+
 ## 0.0.0-alpha.57
 
 - Fix rendering of number attributes in JSX